requirements.txt
flask
flask-sqlalchemy
beautifulsoup4